+++ dcicutils/__init__.py
@@ -1,8 +1,0 @@
-# With this release, you need to import from beanstalk_utils, ff_utils, or s3_utils as appropriate. Don't rely on this.
-# In 4dn-dcic repos, there was only one use in dcicutils.test_base that has been updated.
-# In dbmi-bgm there were no uses.
-#
-# from __future__ import absolute_import
-# from .beanstalk_utils import *  # NOQA
-# from .ff_utils import *  # NOQA
-# from .s3_utils import *  # NOQA